populate_profile_db/requirements.txt
pymongo==4.0.1
tqdm==4.63.0
python-dotenv==0.19.2
